+++ app/src/main/java/com/oreocube/booksearch/MainNavHost.kt
@@ -5,10 +5,10 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
 import androidx.core.net.toUri
 import androidx.navigation.compose.NavHost
-import com.oreocube.booksearch.feature.book.bookDetailScreen
-import com.oreocube.booksearch.feature.book.navigateToBookDetail
-import com.oreocube.booksearch.feature.book.navigateToSearchBook
-import com.oreocube.booksearch.feature.book.searchBookScreen
+import com.oreocube.booksearch.feature.book.detail.bookDetailScreen
+import com.oreocube.booksearch.feature.book.detail.navigateToBookDetail
+import com.oreocube.booksearch.feature.book.search.navigateToSearchBook
+import com.oreocube.booksearch.feature.book.search.searchBookScreen
 import com.oreocube.booksearch.feature.favorite.favoriteLibraryScreen
 import com.oreocube.booksearch.feature.home.HomeRoute
 import com.oreocube.booksearch.feature.home.homeScreen
